app/rag/prompts.py
from langchain_core.prompts import PromptTemplate

RAG_PROMPT_TEMPLATE = """You are an AI assistant specialized in document intelligence.
Your task is to answer the user's question based ONLY on the provided context.

NON-NEGOTIABLE CONSTRAINTS:
1. Answer strictly using the retrieved chunks below.
2. Every factual claim MUST be followed by a citation in the format [ID].
3. If the answer is not contained within the context, say "I don't know based on the provided documents."
4. Do not use any prior knowledge or outside information.
5. Citations must use the chunk_id provided in the metadata.

Context:
{context}

Question: {question}

Answer with citations:"""

RAG_PROMPT = PromptTemplate(
    template=RAG_PROMPT_TEMPLATE,
    input_variables=["context", "question"]
)
