requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
bcrypt==4.0.1
sqlalchemy
psycopg2-binary
pydantic-settings
pydantic[email]
langchain
langchain-community
langchain-classic
langchain-google-genai
pinecone-client
pypdf
python-docx
python-pptx
openpyxl
beautifulsoup4
markdown
python-dotenv
email-validator
jinja2 
passlib 
aiosmtplib